pyproject.toml
[project]
name = "medmnist-conditional-vae"
version = "0.1.0"
description = "Conditional VAE for Medical Imaging using MedMNIST datasets with Beta-VAE and advanced techniques"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "medmnist>=3.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "tqdm>=4.64.0",
    "scikit-learn>=1.3.0",
    "tensorboard>=2.13.0",
    "lpips>=0.1.4",
    "open-clip-torch>=2.20.0",
    "einops>=0.7.0",
    "scipy>=1.13.1",
]

[project.scripts]
train = "main:train"
generate = "generate:main"
evaluate = "evaluate:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
